Use SIGINT instead of SIGTERM to stop the build. Based on a patch from Harald Fernengel. Fixes bug 71132
BUG: 71132

+++ parts/outputviews/makewidget.cpp
@@ -293,7 +293,8 @@ void MakeWidget::startNextJob()
 
 void MakeWidget::killJob()
 {
-	childproc->kill();
+	if (!childproc->kill(SIGINT))
+		childproc->kill();
 }
 
 bool MakeWidget::isRunning()
